refactor: create Menu component for consolidated auth and fix children components
